Updated to Fusions 1.1.0
git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/affxparser@44422 bc3139a8-67e5-0310-9ffc-ced21a209358

+++ src/fusion_sdk/util/md5sum.cpp
@@ -21,16 +21,21 @@
 
 
 //
-#include <assert.h>
+#include <util/md5sum.h>
+//
+#include <util/md5.h>
+//
+#include <cassert>
+#include <cstdlib>
+#include <cstring>
 #include <fstream>
+#include <string>
+//
+
 #ifndef _MSC_VER
 #include <netinet/in.h>
 #include <stdint.h>
 #endif
-#include <stdlib.h>
-//
-#include <util/md5.h>
-#include <util/md5sum.h>
 
 using namespace std;
 
@@ -72,7 +77,7 @@ affx::md5sum::update(const std::string& str)
 int
 affx::md5sum::update(char const * const str)
 {
-  MD5Update(&m_context,(const u_int8_t*)str,strlen(str));
+  MD5Update(&m_context,(const u_int8_t*)str,(u_int32_t)strlen(str));
   return 0;
 }
 
@@ -205,7 +210,7 @@ affx::md5sum::ofFile(const std::string filename,std::string& sum)
 
   do {
     istm.read(buf_ptr,buf_size);
-    update(buf_ptr,istm.gcount());
+    update(buf_ptr,(int)istm.gcount());
   } while (!istm.eof());
 
   delete[] buf_ptr;
